cleanup: move contentValidation&mailsValidation from MessagesController to SendSmsManager
rationale: WsController must not depend on other controllers (needed to use spring MVC instead of jersey)
